Refactor into modular OOP structure
