+++ README.md
@@ -1,1 +1,12 @@
-# Secret-Call-
+# Secret Call
+
+This project provides a one-button WebRTC audio call with automatic
+localhost.run tunnelling. The codebase is now organised into several
+modules to make future extensions easier.
+
+Run the application:
+
+```bash
+python main.py
+```
+
