+++ core.py
@@ -22,13 +22,8 @@ import collections
 import re
 import sys
 import os
-import subprocess
-import pathlib
-import signal
 from dataclasses import dataclass
 from fractions import Fraction
-import webbrowser
-import tkinter as tk
 
 import sounddevice as sd
 import websockets
@@ -65,236 +60,6 @@ logging.basicConfig(
 )
 log = logging.getLogger("SecureCallWebRTC")
 logging.getLogger("aioice").setLevel(logging.WARNING)
-
-# ─── Авто-туннель (localhost.run) ───────────────────────────────────
-_TUNNEL_PROC = None
-_TUNNEL_URL = None
-_TUNNEL_LOCK = threading.Lock()
-_TUNNEL_URL_EVENT = threading.Event()
-
-
-def _which(cmd: str) -> bool:
-    from shutil import which
-    return which(cmd) is not None
-
-def ensure_ssh_key() -> pathlib.Path:
-    """
-    Создаёт ed25519-ключ в ~/.ssh/localhost_run, если его ещё нет.
-    Возвращает путь к приватному ключу.
-    """
-    home = pathlib.Path.home()
-    ssh_dir = home / ".ssh"
-    key_path = ssh_dir / "localhost_run"
-    if not ssh_dir.exists():
-        ssh_dir.mkdir(parents=True, exist_ok=True)
-    if not key_path.exists():
-        log.info("[TUNNEL] Generating SSH key at %s", key_path)
-        # ssh-keygen -t ed25519 -f <path> -N ""
-        result = subprocess.run(
-            ["ssh-keygen", "-t", "ed25519", "-f", str(key_path), "-N", ""],
-            stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True
-        )
-        if result.returncode != 0:
-            raise RuntimeError(f"ssh-keygen failed:\n{result.stdout}")
-    return key_path
-
-def _extract_lhr_https(text: str) -> str | None:
-    """Извлекает первую https-ссылку вида https://<sub>.lhr.life из текста."""
-    m = re.search(r"https://([a-z0-9\-]+\.lhr\.life)(?:[^\s]*)?", text.strip())
-    return f"https://{m.group(1)}" if m else None
-
-def _tunnel_reader(proc: subprocess.Popen, on_url: callable = None):
-    """Читает вывод процесса туннеля и извлекает URL."""
-    global _TUNNEL_URL
-    
-    buf = ""
-    while True:
-        chunk = proc.stdout.read(1) if proc.stdout else None
-        if not chunk:  # Процесс завершился
-            break
-            
-        buf += chunk
-        
-        # Логируем полные строки
-        if chunk == "\n":
-            line = buf.strip()
-            if line:
-                log.info("[TUNNEL] %s", line)
-                
-                # Пытаемся извлечь URL
-                url = _extract_lhr_https(line)
-                if url and not _TUNNEL_URL:
-                    with _TUNNEL_LOCK:
-                        _TUNNEL_URL = url
-                        _TUNNEL_URL_EVENT.set()
-                    log.info("[TUNNEL] URL found: %s", url)
-                    if callable(on_url):
-                        try:
-                            on_url(url)
-                        except Exception as e:
-                            log.warning("[TUNNEL] on_url callback error: %s", e)
-            buf = ""
-        
-        # Параллельно проверяем буфер на наличие URL
-        if not _TUNNEL_URL:
-            url = _extract_lhr_https(buf)
-            if url:
-                with _TUNNEL_LOCK:
-                    _TUNNEL_URL = url
-                    _TUNNEL_URL_EVENT.set()
-                log.info("[TUNNEL] URL found (inline): %s", url)
-                if callable(on_url):
-                    try:
-                        on_url(url)
-                    except Exception as e:
-                        log.warning("[TUNNEL] on_url callback error: %s", e)
-
-def start_localhost_run_tunnel(local_port: int = HTTP_PORT, on_url=None) -> None:
-    """
-    Запускает: ssh -tt -o StrictHostKeyChecking=accept-new -o ServerAliveInterval=30 -R 80:127.0.0.1:<local_port> nokey@localhost.run
-    Без -N, без ключей. Читает баннер побайтно и вытаскивает https://<sub>.lhr.life.
-    """
-    import subprocess, os, threading, time
-    global _TUNNEL_PROC, _TUNNEL_URL
-
-    def _safe_call_cb(url: str):
-        if callable(on_url):
-            try:
-                on_url(url)
-            except Exception as e:
-                log.warning("[TUNNEL] on_url callback error: %s", e)
-
-    with _TUNNEL_LOCK:
-        if _TUNNEL_PROC is not None and _TUNNEL_PROC.poll() is None:
-            log.info("[TUNNEL] already running at %s", _TUNNEL_URL or "<pending>")
-            if _TUNNEL_URL:
-                _safe_call_cb(_TUNNEL_URL)
-            return
-
-        if not _which("ssh"):
-            log.warning("[TUNNEL] OpenSSH 'ssh' not found in PATH=%s — tunnel will not be started", os.getenv("PATH"))
-            return
-
-        host = os.environ.get("LOCALHOST_RUN_HOST", "nokey@localhost.run")
-        cmd = [
-            "ssh",
-            "-tt",                                  # принудительный TTY
-            "-o", "StrictHostKeyChecking=accept-new",
-            "-o", "ServerAliveInterval=30",
-            "-o", "ExitOnForwardFailure=yes",
-            "-R", f"80:127.0.0.1:{local_port}",
-            host,
-            # без -N → интерактивная сессия, баннер полностью приходит
-        ]
-
-        # Важно: не скрываем окно и даём stdin — иначе хвост может не прийти
-        creationflags = 0  # на Windows не используем CREATE_NO_WINDOW
-        env = os.environ.copy()
-        env.setdefault("TERM", "xterm")  # на всякий случай
-
-        log.info("[TUNNEL] Starting localhost.run tunnel (this may take a few seconds)…")
-        _TUNNEL_PROC = subprocess.Popen(
-            cmd,
-            stdin=subprocess.PIPE,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.STDOUT,
-            text=True,
-            creationflags=creationflags,
-            bufsize=1,  # строковое буферизование
-            env=env,
-        )
-
-        def reader():
-            """Логируем баннер побайтно; после 'connection id' отправляем '\\n'; парсим https://<sub>.lhr.life."""
-            global _TUNNEL_URL
-            assert _TUNNEL_PROC.stdout is not None
-            buf = ""
-            nudged = False
-            t0 = time.time()
-            log.info("[TUNNEL] reader started, waiting for https://<sub>.lhr.life …")
-
-            while True:
-                ch = _TUNNEL_PROC.stdout.read(1)
-                if ch == "" or ch is None:
-                    break
-                buf += ch
-
-                if ch == "\n":
-                    line = buf.strip("\r\n")
-                    if line:
-                        log.info("[TUNNEL] %s", line)
-
-                        # «пинок» сразу после connection id
-                        if (not nudged) and ("your connection id is" in line.lower()):
-                            try:
-                                if _TUNNEL_PROC.stdin:
-                                    _TUNNEL_PROC.stdin.write("\n")
-                                    _TUNNEL_PROC.stdin.flush()
-                                    nudged = True
-                                    log.info("[TUNNEL] nudged stdin with newline")
-                            except Exception as e:
-                                log.info("[TUNNEL] nudge failed: %s", e)
-
-                        url = _extract_lhr_https(line)
-                        if url and _TUNNEL_URL is None:
-                            _TUNNEL_URL = url
-                            log.info("[TUNNEL] URL: %s (open this on both devices)", _TUNNEL_URL)
-                            _safe_call_cb(_TUNNEL_URL)
-                    buf = ""
-
-                # Фолбэк: если спустя 3 сек URL нет — ещё один «пинок»
-                if (not nudged) and (_TUNNEL_URL is None) and (time.time() - t0 > 3.0):
-                    try:
-                        if _TUNNEL_PROC.stdin:
-                            _TUNNEL_PROC.stdin.write("\n")
-                            _TUNNEL_PROC.stdin.flush()
-                            nudged = True
-                            log.info("[TUNNEL] timed nudge sent")
-                    except Exception as e:
-                        log.info("[TUNNEL] timed nudge failed: %s", e)
-
-                # Ищем URL прямо в буфере (на случай печати без \n)
-                if _TUNNEL_URL is None:
-                    url_inline = _extract_lhr_https(buf)
-                    if url_inline:
-                        _TUNNEL_URL = url_inline
-                        log.info("[TUNNEL] URL: %s (open this on both devices)", _TUNNEL_URL)
-                        _safe_call_cb(_TUNNEL_URL)
-
-            log.info("[TUNNEL] process ended")
-
-        threading.Thread(target=reader, name="TunnelReader", daemon=True).start()
-
-def stop_localhost_run_tunnel() -> None:
-    """Останавливает SSH-туннель."""
-    global _TUNNEL_PROC, _TUNNEL_URL
-    
-    with _TUNNEL_LOCK:
-        if _TUNNEL_PROC and _TUNNEL_PROC.poll() is None:
-            log.info("[TUNNEL] Stopping tunnel...")
-            try:
-                if os.name == "nt":
-                    _TUNNEL_PROC.send_signal(signal.CTRL_BREAK_EVENT)
-                else:
-                    _TUNNEL_PROC.terminate()
-                
-                _TUNNEL_PROC.wait(timeout=3)
-            except Exception as e:
-                log.warning("[TUNNEL] Error stopping tunnel: %s", e)
-                try:
-                    _TUNNEL_PROC.kill()
-                except Exception:
-                    pass
-        
-        _TUNNEL_PROC = None
-        _TUNNEL_URL = None
-        _TUNNEL_URL_EVENT.clear()
-
-def wait_for_tunnel_url(timeout: float = 30.0) -> str | None:
-    """Ожидает появления URL туннеля."""
-    if _TUNNEL_URL_EVENT.wait(timeout):
-        return _TUNNEL_URL
-    return None
 
 # ─── Утилиты ────────────────────────────────────────────────────────
 def get_local_ip() -> str:
@@ -884,128 +649,3 @@ async def run_peer(ctx: PeerContext):
         player.close()
         await pc.close()
         log.info("[PC] Closed")
-
-# ─── Асинхронный раннер ────────────────────────────────────────────
-class AsyncRunner:
-    def __init__(self):
-        self.loop = None
-        self.thread = None
-        self.ready = threading.Event()
-    def start(self):
-        def target():
-            self.loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(self.loop)
-            self.ready.set()
-            self.loop.run_forever()
-        self.thread = threading.Thread(target=target, name="AsyncLoop", daemon=True)
-        self.thread.start()
-        self.ready.wait()
-    def submit(self, coro):
-        self.ready.wait()
-        return asyncio.run_coroutine_threadsafe(coro, self.loop)
-    def stop(self):
-        if self.loop:
-            self.loop.call_soon_threadsafe(self.loop.stop)
-        if self.thread:
-            self.thread.join(timeout=1)
-
-# ─── GUI «одна кнопка» ─────────────────────────────────────────────
-class App:
-    def __init__(self, root: tk.Tk, runner: AsyncRunner):
-        self.root = root
-        self.runner = runner
-
-        self.root.title("Secure Call — WebRTC (1 кнопка)")
-        wrap = tk.Frame(root)
-        wrap.pack(padx=10, pady=10)
-
-        tk.Label(
-            wrap,
-            text=(
-                "HTTP сервер запущен.\n"
-                "Публичный URL появится ниже, как кликабельная ссылка.\n"
-                f"Локально:  http://{get_local_ip()}:{HTTP_PORT}\n"
-                "(Браузер спросит доступ к микрофону.)"
-            ),
-            justify="left",
-        ).pack()
-
-        # Кликабельная ссылка HTTPS туннеля
-        self.url_label = tk.Label(wrap, text="", fg="blue", cursor="hand2")
-        self.url_label.pack(pady=(6, 10))
-
-        # Кнопки
-        self.start_btn = tk.Button(wrap, text="Подключиться (auto host/join)", command=self.start_call)
-        self.start_btn.pack(pady=10)
-
-        self.stop_btn = tk.Button(wrap, text="Выход", command=self.stop)
-        self.stop_btn.pack()
-
-        self.status = tk.Label(wrap, text="Сервер запущен, туннель стартует…", fg="green")
-        self.status.pack(pady=(8, 0))
-
-    # Показывает ссылку и делает её кликабельной
-    def set_tunnel_url(self, url: str):
-        def _apply():
-            def open_url(_evt):
-                webbrowser.open(url)
-            self.url_label.config(text=url)
-            self.url_label.bind("<Button-1>", open_url)
-            self.set_status("Публичная ссылка готова", "green")
-        self.root.after(0, _apply)
-
-    def start_call(self):
-        self.set_status("Поиск хоста в LAN…", "blue")
-        self.start_btn.config(state="disabled")
-        self.runner.submit(self.auto_host_or_join())
-
-    async def auto_host_or_join(self):
-        info = await udp_discover(timeout=1.0, attempts=3)
-        if info and "host" in info and await wait_port(info["host"], HTTP_PORT, timeout=3.0):
-            host = info["host"]
-            self.set_status(f"Подключаюсь к хосту {host}…", "blue")
-            await run_peer(PeerContext(ws_url=f"ws://{host}:{HTTP_PORT}/ws", is_initiator=True))
-            self.set_status("Звонок завершён", "green")
-            self.start_btn.config(state="normal")
-            return
-
-        self.set_status("Я — ХОСТ. Жду участника…", "green")
-        await run_peer(PeerContext(ws_url=f"ws://127.0.0.1:{HTTP_PORT}/ws", is_initiator=False))
-        self.set_status("Звонок завершён", "green")
-        self.start_btn.config(state="normal")
-
-    def stop(self):
-        self.set_status("Завершение…", "orange")
-        try:
-            stop_localhost_run_tunnel()
-        except Exception:
-            pass
-        time.sleep(0.2)
-        os._exit(0)
-
-    def set_status(self, text: str, color: str = "green"):
-        try:
-            self.root.after(0, lambda: self.status.config(text=text, fg=color))
-        except Exception:
-            self.status.config(text=text, fg=color)
-
-# ─── Точка входа ───────────────────────────────────────────────────
-if __name__ == "__main__":
-    runner = AsyncRunner(); runner.start()
-    runner.submit(start_http_server())
-    start_udp_responder()
-    log.info("[BOOT] HTTP server scheduled, UDP responder started")
-
-    root = tk.Tk()
-    app = App(root, runner)
-
-    # Стартуем туннель и пробрасываем колбэк показа ссылки
-    threading.Thread(
-        target=start_localhost_run_tunnel,
-        name="TunnelStarter",
-        kwargs={"local_port": HTTP_PORT, "on_url": app.set_tunnel_url},
-        daemon=True
-    ).start()
-
-    root.protocol("WM_DELETE_WINDOW", app.stop)
-    root.mainloop()
